+++ src/main/java/kitchenpos/product/dto/ProductRequest.java
@@ -1,25 +1,25 @@
-package kitchenpos.dto;
+package kitchenpos.product.dto;
 
 import java.math.BigDecimal;
-import kitchenpos.domain.Product;
+import kitchenpos.product.domain.Product;
 
-public class ProductResponse {
+public class ProductRequest {
 
     private Long id;
     private String name;
     private BigDecimal price;
 
-    public ProductResponse() {
+    public ProductRequest() {
     }
 
-    public ProductResponse(Long id, String name, BigDecimal price) {
+    public ProductRequest(Long id, String name, BigDecimal price) {
         this.id = id;
         this.name = name;
         this.price = price;
     }
 
-    public static ProductResponse from(Product product) {
-        return new ProductResponse(product.getId(), product.getName(), product.getPrice());
+    public Product toProduct() {
+        return new Product(id, name, price);
     }
 
     public Long getId() {
