+++ src/test/java/kitchenpos/product/ProductFixture.java
@@ -4,7 +4,7 @@ import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import java.math.BigDecimal;
-import kitchenpos.domain.Product;
+import kitchenpos.product.domain.Product;
 import org.springframework.http.MediaType;
 
 public class ProductFixture {
